Add a write up
I did my write up on some trends that I found in the README file. I hope
that's the right place to do it! I also made some changes to the jupyter
notebook itself to help me think about some of the trends - mostly
adding percentages to some of the summary data frames.